Split apart datamodel.ts
This splits datamodel.ts into mainly database and ui focused parts, with
the remainder either being widely shared (and so go to core.ts), or a
for a specific subsystem (types, staging).
