+++ src/datamodel/database.ts
@@ -13,10 +13,18 @@
  * See, the License, for the specific language governing permissions and
  * limitations under the License.
  *
- * Filename: datamodel.ts
+ * Filename: database.ts
  * Description:
  *   TODO
  */
+
+import {ProjectID, ObservationID, RevisionID, DatumID} from './core';
+import {
+  FAIMSConstantCollection,
+  FAIMSTypeCollection,
+  ProjectUIFields,
+  ProjectUIViews,
+} from './typesystem';
 
 export const UI_SPECIFICATION_NAME = 'ui-specification';
 export const PROJECT_SPECIFICATION_PREFIX = 'project-specification';
@@ -65,11 +73,21 @@ export interface NonNullListingsObject extends ListingsObject {
 export interface ActiveDoc {
   _id: string;
   listing_id: string;
-  project_id: string;
+  project_id: ProjectID;
   username: string | null;
   password: string | null;
   friendly_name?: string;
   is_sync: boolean;
+}
+
+/*
+ * Objects that may be contained in a Project's metadata DB
+ */
+
+export interface ProjectPeople {
+  _id: string;
+  _rev?: string; // optional as we may want to include the raw json in places
+  _deleted?: boolean;
 }
 
 /**
@@ -92,40 +110,6 @@ export type ProjectsList = {
   [key: string]: ProjectObject;
 };
 
-/**
- * User readable information about a project
- * Do not use with sync code; UI code only
- */
-export type ProjectID = string;
-
-export interface ProjectInformation {
-  project_id: ProjectID;
-  name: string;
-  description?: string;
-  last_updated?: string;
-  created?: string;
-  status?: string;
-}
-
-/*
- * Objects that may be contained in a Project's metadata DB
- */
-export interface FAIMSType {
-  [key: string]: any; // any for now until we lock down the json
-}
-
-export interface FAIMSTypeCollection {
-  [key: string]: FAIMSType;
-}
-
-export interface FAIMSConstant {
-  [key: string]: any; // any for now until we lock down the json
-}
-
-export interface FAIMSConstantCollection {
-  [key: string]: FAIMSConstant;
-}
-
 export interface ProjectSchema {
   _id?: string; // optional as we may want to include the raw json in places
   _rev?: string; // optional as we may want to include the raw json in places
@@ -133,22 +117,6 @@ export interface ProjectSchema {
   namespace: string;
   constants: FAIMSConstantCollection;
   types: FAIMSTypeCollection;
-}
-
-export interface ProjectUIFields {
-  [key: string]: any;
-}
-
-export interface ProjectUIViews {
-  [key: string]: any;
-}
-
-export interface ProjectUIModel {
-  _id?: string; // optional as we may want to include the raw json in places
-  _rev?: string; // optional as we may want to include the raw json in places
-  fields: ProjectUIFields;
-  views: ProjectUIViews;
-  start_view: string;
 }
 
 export interface EncodedProjectUIModel {
@@ -168,84 +136,6 @@ export interface EncodedProjectMetadata {
   is_attachment: boolean;
   metadata: any;
 }
-
-export interface ProjectPeople {
-  _id: string;
-  _rev?: string; // optional as we may want to include the raw json in places
-  _deleted?: boolean;
-}
-
-// There are two internal ID for observations, the former is unique to a
-// project, the latter unique to the system (i.e. includes project_id)
-export type ObservationID = string;
-export type FullyResolvedObservationID = string;
-export interface SplitObservationID {
-  project_id: string;
-  observation_id: ObservationID;
-}
-export type RevisionID = string;
-export type DatumID = string;
-
-export function resolve_observation_id(
-  split_id: SplitObservationID
-): FullyResolvedObservationID {
-  const cleaned_project_id = split_id.project_id.replace('||', '\\|\\|');
-  return cleaned_project_id + '||' + split_id.observation_id;
-}
-
-export function split_full_observation_id(
-  full_proj_id: FullyResolvedObservationID
-): SplitObservationID {
-  const splitid = full_proj_id.split('||');
-  if (
-    splitid.length !== 2 ||
-    splitid[0].trim() === '' ||
-    splitid[1].trim() === ''
-  ) {
-    throw Error('Not a valid full observation id');
-  }
-  const cleaned_project_id = splitid[0].replace('\\|\\|', '||');
-  return {
-    project_id: cleaned_project_id,
-    observation_id: splitid[1],
-  };
-}
-
-export interface ObservationMetadata {
-  project_id: ProjectID;
-  observation_id: ObservationID;
-  revision_id: RevisionID;
-  created: Date;
-  created_by: string;
-  updated: Date;
-  updated_by: string;
-  conflicts: boolean;
-}
-
-export type ObservationMetadataList = {
-  [key: string]: ObservationMetadata;
-};
-
-// This is used within the form/ui subsystem, do not use with pouch
-export interface Observation {
-  project_id?: ProjectID;
-  observation_id: ObservationID;
-  revision_id: RevisionID | null;
-  type: string;
-  data: {[field_name: string]: any};
-  updated: Date;
-  updated_by: string;
-  /*
-  created{_by} are optional as we don't need to track them with the actual data.
-  If you need creation information, then use observation metadata
-  */
-  created?: Date;
-  created_by?: string;
-}
-
-export type ObservationList = {
-  [key: string]: Observation;
-};
 
 // This is used within the pouch/sync subsystem, do not use with form/ui
 export interface EncodedObservation {
@@ -300,14 +190,6 @@ export interface Datum {
   revision_id: RevisionID;
   observation_id: ObservationID;
   annotations: any;
-}
-
-export interface SavedView {
-  // ID: active_id + '/' + view_name
-  // OR: active_id + '/' + view_name + '/' + existing.observation + '/' + existing.revision
-  _id: string;
-  // Fields
-  [key: string]: unknown;
 }
 
 /*
