+++ src/datamodel/staging.ts
@@ -13,25 +13,15 @@
  * See, the License, for the specific language governing permissions and
  * limitations under the License.
  *
- * Filename: users.ts
+ * Filename: staging.ts
  * Description:
  *   TODO
  */
 
-import {active_db} from './sync/databases';
-import {ProjectID} from './datamodel/core';
-
-export async function getFriendlyUserName(
-  project_id: ProjectID
-): Promise<string> {
-  const doc = await active_db.get(project_id);
-  if (doc.friendly_name === undefined) {
-    return doc.username || 'Dummy User';
-  }
-  return doc.friendly_name;
-}
-
-export async function getCurrentUserId(project_id: ProjectID): Promise<string> {
-  const doc = await active_db.get(project_id);
-  return doc.username || 'Dummy User';
+export interface SavedView {
+  // ID: active_id + '/' + view_name
+  // OR: active_id + '/' + view_name + '/' + existing.observation + '/' + existing.revision
+  _id: string;
+  // Fields
+  [key: string]: unknown;
 }
